:truck: (interface): Rename file interface/CoreGraphics.hpp to interface/Graphics.hpp
